final update with full requirement fullfilled

+++ api/schemas.py
@@ -3,10 +3,10 @@ from pydantic import BaseModel, Field
 class ShiftSafetyRequest(BaseModel):
     corridor_id: str = Field(..., example="Megenagna")
     hour_of_day: int = Field(..., ge=0, le=23, example=22)
-    ambient_light: float = Field(..., ge=0.0, le=1.0, example=0.1)
-    street_light_func: float = Field(..., ge=0.0, le=1.0, example=0.4)
-    crime_risk_index: float = Field(..., ge=0.0, le=10.0, example=6.5)
-    foot_traffic: float = Field(..., ge=0.0, le=100.0, example=15.0)
+    ambient_light: int = Field(..., ge=0.0, le=1.0, example=0.1)
+    street_light_func: int = Field(..., ge=0.0, le=1.0, example=0.4)
+    crime_risk_index: int = Field(..., ge=0.0, le=10.0, example=6.5)
+    foot_traffic: int = Field(..., ge=0.0, le=100.0, example=15.0)
 
 class ShiftSafetyResponse(BaseModel):
     corridor_id: str = Field(..., json_schema_extra={"example": "Megenagna"})
